Take uint256 penalty parameter in reallocate and allocateFromIdle
Per review suggestions; the config comparison widens implicitly and
setPenalty keeps its uint64 parameter.

+++ src/periphery/blue-public-allocator/BluePublicAllocator.sol
@@ -118,7 +118,7 @@ contract BluePublicAllocator is IBluePublicAllocator {
         address allocateAdapter,
         MarketParams calldata allocateMarketParams,
         uint128 assets,
-        uint64 penalty
+        uint256 penalty
     ) external {
         require(vaultData[vault].penalty == penalty, IncorrectPenalty());
         uint256 penaltyAssets = MathLib.mulDivUp(assets, penalty, WAD);
@@ -146,7 +146,7 @@ contract BluePublicAllocator is IBluePublicAllocator {
         address adapter,
         MarketParams calldata marketParams,
         uint128 assets,
-        uint64 penalty
+        uint256 penalty
     ) external {
         require(vaultData[vault].penalty == penalty, IncorrectPenalty());
         uint256 penaltyAssets = MathLib.mulDivUp(assets, penalty, WAD);
